+++ lib/resources.go
@@ -32,17 +32,22 @@ func GetResources(stackname *string, svc interface {
 	if *stackname != "" && !strings.Contains(*stackname, "*") {
 		input.StackName = stackname
 	}
-	resp, err := svc.DescribeStacks(context.TODO(), input)
-	if err != nil {
-		var bne *smithy.OperationError
-		if errors.As(err, &bne) {
-			log.Fatalln("error:", bne.Err)
+	paginator := cloudformation.NewDescribeStacksPaginator(svc, input)
+	allstacks := make([]types.Stack, 0)
+	for paginator.HasMorePages() {
+		output, err := paginator.NextPage(context.TODO())
+		if err != nil {
+			var bne *smithy.OperationError
+			if errors.As(err, &bne) {
+				log.Fatalln("error:", bne.Err)
+			}
+			log.Fatalln(err)
 		}
-		log.Fatalln(err)
+		allstacks = append(allstacks, output.Stacks...)
 	}
 	stackRegex := "^" + strings.ReplaceAll(*stackname, "*", ".*") + "$"
 	tocheckstacks := make([]types.Stack, 0)
-	for _, stack := range resp.Stacks {
+	for _, stack := range allstacks {
 		if strings.Contains(*stackname, "*") {
 			if matched, _ := regexp.MatchString(stackRegex, *stack.StackName); !matched {
 				continue